// src/app/scenes/phoenixFlame/PhoenixFlame.types.ts
export type FlameParticleState = {
  active: boolean;
  spark: boolean;
  x: number;
  y: number;
  rotation: number;
  width: number;
  height: number;
  alpha: number;
  tint: number;
  z: number;
  ageMs: number;
  lifeMs: number;
  vx: number;
  vy: number;
  startSize: number;
  endSize: number;
  spin: number;
};